Create Energy transfer in blockchain
In this simple contract, there is only one energy producer and one energy consumer. The contract is initialized with the addresses of the producer and consumer, the amount of energy to be transferred, and the price per unit of energy. The buyEnergy function is called by the consumer to transfer the energy and the corresponding payment to the producer. The cancelTransaction function is called by the producer to cancel the transaction and receive a refund of any payment that has been made